+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - [#136 - Charts cards should specify mode](https://github.com/dynatrace-extensions/dynatrace-extensions-vscode/issues/136)
 - [#132 - Error "cannot read name" when activating new extension](https://github.com/dynatrace-extensions/dynatrace-extensions-vscode/issues/132)
 - [#135 - Screens rejected on subsequent uploads](https://github.com/dynatrace-extensions/dynatrace-extensions-vscode/issues/135)
+- [#127 - .DS_Store files may impact zip archiving for Mac users](https://github.com/dynatrace-extensions/dynatrace-extensions-vscode/issues/127)
 
 ## Version 2.2.0 (27.07.2023)